backend/api/Projetcs-service/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pydantic==2.0.0
pydantic-settings==2.0.0